# Conflicts:
#	api-demo/pom.xml
#	api-demo/src/main/java/com/jk/model/MyOrder.java

+++ api-demo/src/main/java/com/jk/model/Book.java
@@ -45,7 +45,7 @@ public class Book implements Serializable {
 
     private Integer bookCount;
 
-    private Double hits;
+    private Integer hits;
 
     private Integer bookStar;
